requirements.txt
llama2-wrapper
langchain
langchain-openai
langchain_pinecone
langchain[docarray]
docarray
pydantic==1.10.8
pytube 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
pypdf
langchain-community
faiss-gpu
chromadb==0.5.3
pytest
black
pylint